settings moved to separate file
