+++ src/config/settings.js
@@ -1,0 +1,14 @@
+let settings = {
+  set: false,
+  main_vehicle: 'car',
+  main_vehicle_id: -1,
+  vehicles_selected: {
+    car: false,
+    truck: false,
+    bus: false
+  }
+};
+
+module.exports = {
+  settings: settings,
+};
